Added section header def

+++ src/data_directory.hpp
@@ -1,11 +1,12 @@
 #pragma once
 #include <cstdint>
+#include <cstddef>
 
 namespace pe
 {
 	struct data_directory
 	{
-		static constexpr std::uint32_t count = 16;
+		static constexpr std::size_t count = 16;
 
 		std::uint32_t virtual_address;
 		std::uint32_t size;
